requirements.txt
wheel
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-cleanup==4.0.0
django-tinymce4-lite==1.8.0
idna==2.9
jsmin==2.2.2
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
google-api-core==1.22.0
google-auth==1.20.1
google-cloud-secret-manager==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
google-cloud-storage
boto3
django-cors-headers
tenacity